// backend/src/core/interfaces/controllers/i-squad-controller.ts
import type { RequestHandler } from "express";

export interface ISquadController {
  createSquad: RequestHandler;
  getAllSquads: RequestHandler;
  toggleSquad: RequestHandler;
  getSquadsByCategory: RequestHandler;
  joinSquad: RequestHandler;
  getSquadDetailsByHandle: RequestHandler;
  getSquadContents: RequestHandler;
  leaveSquad: RequestHandler;
  getJoinedSquads: RequestHandler;
}